requirements.txt
absl-py==2.1.0 ; python_version >= "3.10" and python_version < "4.0"
aiohappyeyeballs==2.3.6 ; python_version >= "3.10" and python_version < "4.0"
aiohttp==3.10.3 ; python_version >= "3.10" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
async-timeout==4.0.3 ; python_version >= "3.10" and python_version < "3.11"
attrs==24.2.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2024.7.4 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
datasets==2.21.0 ; python_version >= "3.10" and python_version < "4.0"
deepspeed==0.14.5 ; python_version >= "3.10" and python_version < "4.0"
dill==0.3.8 ; python_version >= "3.10" and python_version < "4.0"
docker-pycreds==0.4.0 ; python_version >= "3.10" and python_version < "4.0"
filelock==3.15.4 ; python_version >= "3.10" and python_version < "4.0"
frozenlist==1.4.1 ; python_version >= "3.10" and python_version < "4.0"
fsspec==2024.6.1 ; python_version >= "3.10" and python_version < "4.0"
fsspec[http]==2024.6.1 ; python_version >= "3.10" and python_version < "4.0"
gitdb==4.0.11 ; python_version >= "3.10" and python_version < "4.0"
gitpython==3.1.43 ; python_version >= "3.10" and python_version < "4.0"
grpcio==1.65.5 ; python_version >= "3.10" and python_version < "4.0"
hjson==3.1.0 ; python_version >= "3.10" and python_version < "4.0"
huggingface-hub==0.24.5 ; python_version >= "3.10" and python_version < "4.0"
idna==3.7 ; python_version >= "3.10" and python_version < "4.0"
joblib==1.4.2 ; python_version >= "3.10" and python_version < "4.0"
markdown==3.7 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.10" and python_version < "4.0"
mecab-python3==1.0.9 ; python_version >= "3.10" and python_version < "4.0"
multidict==6.0.5 ; python_version >= "3.10" and python_version < "4.0"
multiprocess==0.70.16 ; python_version >= "3.10" and python_version < "4.0"
ninja==1.11.1.1 ; python_version >= "3.10" and python_version < "4.0"
nltk==3.8.1 ; python_version >= "3.10" and python_version < "4.0"
numpy==1.26.4 ; python_version >= "3.10" and python_version < "4.0"
nvidia-cublas-cu11==11.10.3.66 ; python_version >= "3.10" and python_version < "4.0"
nvidia-cuda-nvrtc-cu11==11.7.99 ; python_version >= "3.10" and python_version < "4.0"
nvidia-cuda-runtime-cu11==11.7.99 ; python_version >= "3.10" and python_version < "4.0"
nvidia-cudnn-cu11==8.5.0.96 ; python_version >= "3.10" and python_version < "4.0"
nvidia-ml-py==12.560.30 ; python_version >= "3.10" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.10" and python_version < "4.0"
pandas==2.2.2 ; python_version >= "3.10" and python_version < "4.0"
pathtools==0.1.2 ; python_version >= "3.10" and python_version < "4.0"
pillow==10.4.0 ; python_version >= "3.10" and python_version < "4.0"
promise==2.3 ; python_version >= "3.10" and python_version < "4.0"
protobuf==4.25.4 ; python_version >= "3.10" and python_version < "4.0"
psutil==6.0.0 ; python_version >= "3.10" and python_version < "4.0"
py-cpuinfo==9.0.0 ; python_version >= "3.10" and python_version < "4.0"
pyarrow==17.0.0 ; python_version >= "3.10" and python_version < "4.0"
pydantic-core==2.20.1 ; python_version >= "3.10" and python_version < "4.0"
pydantic==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
pydeprecate==0.3.2 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "4.0"
pytorch-lightning==1.7.7 ; python_version >= "3.10" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.10" and python_version < "4.0"
regex==2024.7.24 ; python_version >= "3.10" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.10" and python_version < "4.0"
rouge-metric==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
sentry-sdk==2.13.0 ; python_version >= "3.10" and python_version < "4.0"
setproctitle==1.3.3 ; python_version >= "3.10" and python_version < "4.0"
setuptools==72.2.0 ; python_version >= "3.10" and python_version < "4.0"
shortuuid==1.0.13 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
smmap==5.0.1 ; python_version >= "3.10" and python_version < "4.0"
tensorboard-data-server==0.7.2 ; python_version >= "3.10" and python_version < "4.0"
tensorboard==2.17.1 ; python_version >= "3.10" and python_version < "4.0"
tokenizers==0.13.3 ; python_version >= "3.10" and python_version < "4.0"
torch==1.13.0 ; python_version >= "3.10" and python_version < "4.0"
torchmetrics==0.9.3 ; python_version >= "3.10" and python_version < "4.0"
tqdm==4.66.5 ; python_version >= "3.10" and python_version < "4.0"
transformers==4.23.1 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.2.2 ; python_version >= "3.10" and python_version < "4.0"
wandb==0.13.4 ; python_version >= "3.10" and python_version < "4.0"
werkzeug==3.0.3 ; python_version >= "3.10" and python_version < "4.0"
wheel==0.44.0 ; python_version >= "3.10" and python_version < "4.0"
xxhash==3.5.0 ; python_version >= "3.10" and python_version < "4.0"
yarl==1.9.4 ; python_version >= "3.10" and python_version < "4.0"
